Refactor hash table implementation to support long long values; update related functions and documentation

+++ src/hash_w.h
@@ -1,3 +1,12 @@
+/**
+ * Author:  Zhean Robby Ganituen
+ * Date:    December 19, 2024
+ * Project: zlang Programming Language
+ * 
+ * Contains the hash for variables, function names, etc.
+ * Wrapper for the C++ class.
+ */
+
 #ifndef HASH_W
 #define HASH_W
 
@@ -7,8 +16,8 @@ extern "C"{
 
 void* __HASH_VARIABLES_MAKE__();
 void __HASH_VARIABLES_DELETE__(void *hash);
-void __HASH_VARIABLES_PUT__(void *hash, const char *key, int value);
-int __HASH_VARIABLES_GET__(void *hash, const char *key);
+void __HASH_VARIABLES_PUT__(void *hash, const char *key, long long value);
+long long __HASH_VARIABLES_GET__(void *hash, const char *key);
 
 #ifdef __cplusplus
 }
